changed all occurances of 'float' for money to 'decimal'

+++ SpendingTracker/Managers/DayManager.cs
@@ -3,5 +3,5 @@
 public static class DayManager
 {
     // TODO: would be nice to load it from file, so we can change it in app.
-    public static readonly float DAY_INITIAL_BUDGET = 100;
+    public static readonly decimal DAY_INITIAL_BUDGET = 100;
 }
