+++ SpendingTracker/Models/SpendingTransaction.cs
@@ -3,5 +3,5 @@
 public class SpendingTransaction
 {
     // TODO: add spending name for extended statistics
-    public float SpentValue { get; set; }
+    public decimal SpentValue { get; set; }
 }
